pyproject.toml
[project]
name = "agent-runner"
version = "1.0.0"
description = "AI-powered agent runner using OpenHands SDK"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "httpx>=0.27.0",
    "pydantic>=2.9.0",
]

[project.optional-dependencies]
server = [
    "fastapi>=0.115.0",
    "uvicorn>=0.32.0",
]
openhands = [
    "openhands-sdk; python_version >= '3.12'",
    "openhands-tools; python_version >= '3.12'",
]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
]

[project.scripts]
agent-runner = "agent_runner.cli:main"
agent-runner-server = "server.app:run_server"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/agent_runner", "server"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
